move config as different package
